ml_service/requirements.txt
flask==3.0.3
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
werkzeug>=3.0.0
rembg>=2.0.0

